unified interface to use from ifigure.interactive import *, as a starting point to plotting with piScope
