+++ ChangeLog
@@ -1,3 +1,9 @@
+2026 8/30
+	* Unified plotting entry point: ifigure.interactive
+	** added internal backends: ifigure._interactive_gui and ifigure._interactive_nogui
+	** changed ifigure.client to a compatibility shim that re-exports no-GUI backend
+	** retained backward compatibility for existing user scripts
+
 2026 8/29
 	* ScriptEditor now infers syntax mode from file extension when opening files.
 	* Unknown or extension-less files now fall back to Syntax=None.
