+++ python/ifigure/__init__.py
@@ -1,3 +1,4 @@
 __version__ = '26.8.29'
 
 from .piscope import piscope
+from .llmrun import llmrun
