+++ python/ifigure/_private/__init__.py
@@ -1,0 +1,1 @@
+"""Internal private modules for ifigure."""
